requirements.txt
Django>=5.2.7
djangorestframework>=3.15.2
djangorestframework-simplejwt>=5.3.1
django-environ>=0.11.2
Pillow>=10.4.0
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
drf-yasg>=1.21.7
django-filter>=23.5